HV: treewide: unify the type of bit-field members
Qualified or unqualified int or bool are the only types allowed for bit-field
members in C99, and MISRA C further forbids using plain int.

Use uint32_t (which is equivalent to unsigned int) for all bit-field members.

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/arch/x86/lapic.c
@@ -20,26 +20,26 @@ union apic_icr {
 		uint32_t hi_32;
 	} value_32;
 	struct {
-		uint64_t vector:8;
-		uint64_t delivery_mode:3;
-		uint64_t destination_mode:1;
-		uint64_t delivery_status:1;
-		uint64_t rsvd_1:1;
-		uint64_t level:1;
-		uint64_t trigger_mode:1;
-		uint64_t rsvd_2:2;
-		uint64_t shorthand:2;
-		uint64_t rsvd_3:12;
-		uint64_t rsvd_4:32;
+		uint32_t vector:8;
+		uint32_t delivery_mode:3;
+		uint32_t destination_mode:1;
+		uint32_t delivery_status:1;
+		uint32_t rsvd_1:1;
+		uint32_t level:1;
+		uint32_t trigger_mode:1;
+		uint32_t rsvd_2:2;
+		uint32_t shorthand:2;
+		uint32_t rsvd_3:12;
+		uint32_t rsvd_4:32;
 	} bits;
 	struct {
-		uint64_t rsvd_1:32;
-		uint64_t rsvd_2:24;
-		uint64_t dest_field:8;
+		uint32_t rsvd_1:32;
+		uint32_t rsvd_2:24;
+		uint32_t dest_field:8;
 	} x_bits;
 	struct {
-		uint64_t rsvd_1:32;
-		uint64_t dest_field:32;
+		uint32_t rsvd_1:32;
+		uint32_t dest_field:32;
 	} x2_bits;
 };
 
@@ -116,13 +116,13 @@ union apic_lvt {
 union lapic_base_msr {
 	uint64_t value;
 	struct {
-		uint64_t rsvd_1:8;
-		uint64_t bsp:1;
-		uint64_t rsvd_2:1;
-		uint64_t x2APIC_enable:1;
-		uint64_t xAPIC_enable:1;
-		uint64_t lapic_paddr:24;
-		uint64_t rsvd_3:28;
+		uint32_t rsvd_1:8;
+		uint32_t bsp:1;
+		uint32_t rsvd_2:1;
+		uint32_t x2APIC_enable:1;
+		uint32_t xAPIC_enable:1;
+		uint32_t lapic_paddr:24;
+		uint32_t rsvd_3:28;
 	} fields;
 };
